app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.19
botocore==1.35.19
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.2
h11==0.14.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
numpy==2.1.1
nvidia-nccl-cu12==2.23.4
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
xgboost==2.1.1
